+++ app/src/main/java/com/sesac/developer_study_platform/ui/mypage/MyPageFragment.kt
@@ -11,7 +11,7 @@ import androidx.navigation.fragment.findNavController
 import com.sesac.developer_study_platform.EventObserver
 import com.sesac.developer_study_platform.R
 import com.sesac.developer_study_platform.databinding.FragmentMyPageBinding
-import com.sesac.developer_study_platform.isNetworkConnected
+import com.sesac.developer_study_platform.util.isNetworkConnected
 import com.sesac.developer_study_platform.ui.common.SpaceItemDecoration
 import com.sesac.developer_study_platform.ui.common.StudyAdapter
 import com.sesac.developer_study_platform.ui.common.StudyClickListener
@@ -49,7 +49,7 @@ class MyPageFragment : Fragment() {
         setBookmarkButton()
         setDialogButton()
         setNavigation()
-        networkStatus()
+        binding.isNetworkConnected = isNetworkConnected(requireContext())
     }
 
     private fun setStudyAdapter() {
@@ -148,14 +148,6 @@ class MyPageFragment : Fragment() {
                 findNavController().navigate(action)
             }
         )
-    }
-
-    private fun networkStatus() {
-        if (!isNetworkConnected(requireContext())) {
-            binding.networkStatus.visibility = View.VISIBLE
-        } else {
-            binding.networkStatus.visibility = View.GONE
-        }
     }
 
     override fun onDestroyView() {
