+++ app/src/main/java/com/sesac/developer_study_platform/ui/profile/ProfileFragment.kt
@@ -8,12 +8,12 @@ import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
-import com.sesac.developer_study_platform.EventObserver
 import androidx.navigation.fragment.navArgs
+import com.sesac.developer_study_platform.EventObserver
 import com.sesac.developer_study_platform.R
 import com.sesac.developer_study_platform.databinding.FragmentProfileBinding
-import com.sesac.developer_study_platform.isNetworkConnected
 import com.sesac.developer_study_platform.ui.common.SpaceItemDecoration
+import com.sesac.developer_study_platform.util.isNetworkConnected
 
 class ProfileFragment : Fragment() {
 
@@ -40,7 +40,7 @@ class ProfileFragment : Fragment() {
         loadUser()
         loadRepositoryList()
         setNavigation()
-        networkStatus()
+        binding.isNetworkConnected = isNetworkConnected(requireContext())
     }
 
     private fun setBackButton() {
@@ -112,14 +112,6 @@ class ProfileFragment : Fragment() {
                 findNavController().navigate(action)
             }
         )
-    }
-
-    private fun networkStatus() {
-        if (!isNetworkConnected(requireContext())) {
-            binding.networkStatus.visibility = View.VISIBLE
-        } else {
-            binding.networkStatus.visibility = View.GONE
-        }
     }
 
     override fun onDestroyView() {
