+++ app/src/main/java/com/sesac/developer_study_platform/ui/searchcategory/SearchCategoryFragment.kt
@@ -4,12 +4,14 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.navArgs
 import com.google.android.material.tabs.TabLayoutMediator
 import com.sesac.developer_study_platform.Category
+import com.sesac.developer_study_platform.R
 import com.sesac.developer_study_platform.databinding.FragmentSearchCategoryBinding
-import com.sesac.developer_study_platform.isNetworkConnected
+import com.sesac.developer_study_platform.util.isNetworkConnected
 
 class SearchCategoryFragment : Fragment() {
 
@@ -22,7 +24,7 @@ class SearchCategoryFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentSearchCategoryBinding.inflate(inflater, container, false)
+        _binding = DataBindingUtil.inflate(inflater, R.layout.fragment_search_category, container, false)
         return binding.root
     }
 
@@ -33,7 +35,7 @@ class SearchCategoryFragment : Fragment() {
         binding.tl.post {
             binding.tl.getTabAt(args.position)?.select()
         }
-        networkStatus()
+        binding.isNetworkConnected = isNetworkConnected(requireContext())
     }
 
     private fun setViewPager() {
@@ -45,14 +47,6 @@ class SearchCategoryFragment : Fragment() {
         TabLayoutMediator(binding.tl, binding.vp) { tab, position ->
             tab.text = getString(Category.entries[position].resId)
         }.attach()
-    }
-
-    private fun networkStatus() {
-        if (!isNetworkConnected(requireContext())) {
-            binding.networkStatus.visibility = View.VISIBLE
-        } else {
-            binding.networkStatus.visibility = View.GONE
-        }
     }
 
     override fun onDestroyView() {
